requirements.txt
langchain_core==0.3.60
langchain_openai==0.3.17
numpy==1.25.2
pandas==2.0.3
pydantic==2.11.4
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
tabulate==0.9.0
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
